convert dynamic to external
