+++ extensions/android-external-audio-processing/BUILD.gn
@@ -12,9 +12,6 @@ rtc_library("external_processing") {
         "external_processing.hpp",
         "external_processing.cc",
         "external_processing_factory_jni.cc",
-        "dynamic_processing.hpp",
-        "dynamic_processing.cc",
-        "dynamic_processing_factory_jni.cc",
     ]
     lib_dirs = []
     deps = [
@@ -29,16 +26,13 @@ rtc_library("external_processing") {
 rtc_android_library("external_java") {
     visibility = [ "*" ]
     sources = [
-        "java/src/org/webrtc/DynamicAudioProcessingFactory.java",
         "java/src/org/webrtc/ExternalAudioProcessingFactory.java",
-        "java/src/org/webrtc/ExternalAudioProcessorFactory.java",
     ]
     deps = ["//sdk/android:peerconnection_java",]
 }
 
 generate_jni("generated_external_jni") {
     sources = [
-        "java/src/org/webrtc/DynamicAudioProcessingFactory.java",
         "java/src/org/webrtc/ExternalAudioProcessingFactory.java",
     ]
     namespace = "external"
