+++ extensions/android-external-audio-processing/external_processing.hpp
@@ -1,9 +1,8 @@
-#ifndef EXTERNAL_PROCESSING_HPP
-#define EXTERNAL_PROCESSING_HPP
+#ifndef DYNAMIC_PROCESSING_HPP
+#define DYNAMIC_PROCESSING_HPP
 
 #include <syslog.h>
 
-#include "include/external_processor.hpp"
 #include "modules/audio_processing/audio_buffer.h"
 #include "modules/audio_processing/audio_processing_impl.h"
 #include "modules/audio_processing/include/audio_processing.h"
@@ -18,12 +17,10 @@ class ExternalProcessing : public webrtc::CustomProcessing {
   ExternalProcessing& operator=(ExternalProcessing&&) = delete;
   ~ExternalProcessing();
 
-  static ExternalProcessing* getInstance(ExternalProcessor* processor) {
-    ::syslog(LOG_INFO, "EXTERNAL-HPP: ExternalProcessing #getInstance; processor: %p", (void*)processor);
+  static ExternalProcessing* getInstance(const char* libname_cstr) {
     if (m_instance == nullptr) {
-      m_instance = new ExternalProcessing(processor);
+      m_instance = new ExternalProcessing(libname_cstr);
     }
-    ::syslog(LOG_INFO, "EXTERNAL-HPP: ExternalProcessing #getInstance; m_instance: %p", (void*)m_instance);
     return m_instance;
   }
 
@@ -34,12 +31,12 @@ class ExternalProcessing : public webrtc::CustomProcessing {
       webrtc::AudioProcessing::RuntimeSetting setting) override;
 
  private:
-  ExternalProcessing(ExternalProcessor* processor);
+  ExternalProcessing(const char* libname_cstr);
 
-  ExternalProcessor* external_processor;
+  const char* libname_cstr;
 
   static ExternalProcessing* m_instance;
 };
 }  // namespace external
 
-#endif  // EXTERNAL_PROCESSING_HPP
+#endif  // DYNAMIC_PROCESSING_HPP
