+++ extensions/android-external-audio-processing/include/external_processor.hpp
@@ -10,9 +10,9 @@ class ExternalProcessor {
  public:
   // Initializes the processor with a specific sample rate and number of
   // channels.
-  virtual void Init(int sample_rate_hz, int num_channels) = 0;
+  virtual int Init(int sample_rate_hz, int num_channels) = 0;
   // Processes the audio data.
-  virtual void ProcessFrame(float* const* channels,
+  virtual int ProcessFrame(float* const* channels,
                             size_t num_frames,
                             size_t num_bands,
                             size_t num_channels) = 0;
