--- extensions/android-external-audio-processing/java/src/org/webrtc/DynamicAudioProcessingFactory.java
@@ -1,28 +1,0 @@
-package org.webrtc;
-
-import org.webrtc.AudioProcessingFactory;
-
-public class DynamicAudioProcessingFactory implements AudioProcessingFactory {
-
-  private final String libname;
-
-  public DynamicAudioProcessingFactory(String libname) {
-    if (libname == null) {
-      throw new NullPointerException("libname must not be null.");
-    }
-    if (libname.isEmpty()) {
-      throw new IllegalArgumentException("libname must not be empty.");
-    }
-    this.libname = libname;
-  }
-
-  @Override
-  public long createNative() {
-    return nativeGetApm(libname);
-  }
-
-  private static native long nativeGetApm(String libname);
-
-}
-
-
