--- extensions/android-external-audio-processing/java/src/org/webrtc/ExternalAudioProcessorFactory.java
@@ -1,11 +1,0 @@
-package org.webrtc;
-
-import org.webrtc.AudioProcessingFactory;
-
-public interface ExternalAudioProcessorFactory {
-
-  public long createNative();
-
-}
-
-
